chore: :wrench: Setting Side Bar -> Atomic Component로 변경
- 여러 곳에서 사용되어 아토믹한 컴포넌트로 변경
