+++ FE/issue-tracker/src/Components/AtomicComponents/SettingSideBar/PlusButton.tsx
@@ -1,7 +1,8 @@
 import { useSetRecoilState } from "recoil";
-import { showDropDownState } from "@/Components/SettingSideBar/SettingSideBarStore";
+import { showDropDownState } from "@/Components/AtomicComponents/SettingSideBar/SettingSideBarStore";
 import IconButton from "@material-ui/core/IconButton";
 import AddIcon from "@material-ui/icons/Add";
+import React from "react";
 
 interface Props {
   id: string;
@@ -12,10 +13,12 @@ const PlusButton = ({ id }: Props) => {
   const setLabelShowFlag = useSetRecoilState(showDropDownState.label);
   const setMileStoneShowFlag = useSetRecoilState(showDropDownState.mileStone);
 
-  const handleOnClick = (e: any) => {
-    const clicked = e.target.closest(`#${id}`).id;
-    if (clicked === "담당자") setAssigneeShowFlag((prev) => !prev);
-    else if (clicked === "레이블") setLabelShowFlag((prev) => !prev);
+  const handleOnClick = (e: React.MouseEvent<HTMLButtonElement>) => {
+    const clicked = (e.target as HTMLElement).closest(`#${id}`);
+    if ((clicked as HTMLElement).id === "담당자")
+      setAssigneeShowFlag((prev) => !prev);
+    else if ((clicked as HTMLElement).id === "레이블")
+      setLabelShowFlag((prev) => !prev);
     else setMileStoneShowFlag((prev) => !prev);
   };
