+++ FE/issue-tracker/src/Components/AtomicComponents/SettingSideBar/SettingDropDown/Item.tsx
@@ -1,6 +1,6 @@
 import MyIcon from "./MyIcon";
 import ItemCheckBox from "./ItemCheckBox";
-import { SettingSideBar as S } from "@/Components/SettingSideBar/SettingSideBarStyles";
+import { SettingSideBar as S } from "@/Components/AtomicComponents/SettingSideBar/SettingSideBarStyles";
 
 interface Props {
   id: string;
