+++ FE/issue-tracker/src/Components/AtomicComponents/SettingSideBar/SettingDropDown/ItemCheckBox.tsx
@@ -1,0 +1,7 @@
+import { SettingSideBar as S } from "@/Components/AtomicComponents/SettingSideBar/SettingSideBarStyles";
+
+const ItemCheckBox = () => {
+  return <S.ItemCheckBox color="primary" />;
+};
+
+export default ItemCheckBox;
