+++ FE/issue-tracker/src/Components/AtomicComponents/SettingSideBar/SettingDropDown/MyIcon.tsx
@@ -1,0 +1,11 @@
+import { SettingSideBar as S } from "@/Components/AtomicComponents/SettingSideBar/SettingSideBarStyles";
+
+interface Props {
+  id: string;
+}
+
+const MyIcon = ({ id }: Props) => {
+  return id === "담당자" ? <S.UserImage /> : <S.Label />;
+};
+
+export default MyIcon;
