+++ FE/issue-tracker/src/Components/AtomicComponents/SettingSideBar/SettingSideBarStore.tsx
@@ -19,3 +19,5 @@ export const showDropDownState = {
     default: false,
   }),
 };
+
+// 담당자 , 레이블 , 마일스톤을 받
