--- FE/issue-tracker/src/Components/SettingSideBar/SettingDropDown/ItemCheckBox.tsx
@@ -1,12 +1,0 @@
-import { StylesProvider } from "@material-ui/core";
-import { SettingSideBar as S } from "@/Components/SettingSideBar/SettingSideBarStyles";
-
-const ItemCheckBox = () => {
-  return (
-    <StylesProvider injectFirst>
-      <S.ItemCheckBox color="primary" />
-    </StylesProvider>
-  );
-};
-
-export default ItemCheckBox;
